just update tam_tools and how test run
