+++ run_tests.py
@@ -1,4 +1,4 @@
-import unittest
-from tamcolors.tests.tam.tests import*
-from tamcolors.tests.tam_tools.tests import*
-unittest.main()
+from tamcolors.tests.all_tests import tests_main
+
+if __name__ == "__main__":
+    tests_main()
