+++ tamcolors/tests/tam_tools/__init__.py
@@ -1,3 +1,22 @@
-from . import tests
+from . import tam_print_tests
+from . import tam_film_tests
+from . import tam_color_palette_tests
+from . import tam_key_manager_tests
+from . import tam_str_tests
+from . import tam_test_box_tests
+from . import tam_placing_tests
+from . import tam_fade_tests
+from . import tam_menu_tests
+from . import tam_list_buffer_tests
 
-__all__ = ("tests",)
+
+__all__ = ("tam_print_tests",
+           "tam_film_tests",
+           "tam_color_palette_tests",
+           "tam_key_manager_tests",
+           "tam_str_tests",
+           "tam_test_box_tests",
+           "tam_placing_tests",
+           "tam_fade_tests",
+           "tam_menu_tests",
+           "tam_list_buffer_tests")
